requirements-dev.txt
fastapi
uvicorn[standard]
pydantic
passlib[bcrypt]
python-jose[cryptography]
pytest
httpx
flake8
black
